Add flly worked completion

+++ src/editor/completion/get_name_from.c
@@ -6,7 +6,7 @@
 /*   By: aviscogl <aviscogl@student.le-101.fr>      +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2018/02/01 10:23:39 by aviscogl     #+#   ##    ##    #+#       */
-/*   Updated: 2018/02/01 18:28:47 by aviscogl    ###    #+. /#+    ###.fr     */
+/*   Updated: 2018/02/03 09:18:16 by aviscogl    ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -16,16 +16,16 @@
 char	*get_name_from(char *str)
 {
 	const size_t	len = ft_strlen(str);
-	int				i;
-	int				prev;
+	size_t			i;
+	size_t			prev;
 
 	i = 0;
 	prev = 0;
-	while (i < (int)len)
+	while (i < len)
 	{
 		if (str[i] == '/')
 			prev = i;
 		i++;
 	}
-	return (ft_strsub(str, prev + 1, len - (prev + 1)));
+	return (ft_strsub(str, (unsigned int)prev + 1, len - (prev + 1)));
 }
